Seq Long으로 변경

+++ src/main/java/com/education/onlinecampus/data/entity/Course.java
@@ -20,7 +20,7 @@ public class Course implements EntityMarker {
     @Column(
             nullable = false
     )
-    private Integer courseSeq;
+    private Long courseSeq;
 
     /** 강좌명 */
     @Column(
